dom/devicestorage/test/test_add.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=786922
-->
<head>
  <title>Test for basic sanity of the device storage API </title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="devicestorage_common.js"></script>

<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=786922">Mozilla Bug 786922</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script class="testbody" type="text/javascript">


devicestorage_setup(function () {

function add(storage, mime) {
  dump("adding: " + mime + "\n");
  return navigator.getDeviceStorage(storage).add(createRandomBlob(mime));
}

var tests = [
  function () { return add("pictures", "image/png")},
  function () { return add("videos",   "video/webm")},
  function () { return add("music",    "audio/wav")},
  function () { return add("sdcard",   "maple/cookies")},
];

function fail(e) {
  ok(false, "onerror was called");
  SimpleTest.finish();
}

function next(e) {

  if (e != undefined)
    ok(true, "addError was called");

  var f = tests.pop();

  if (f == undefined) {
    SimpleTest.finish();
    return;
  }

  request = f();
  request.onsuccess = next;
  request.onerror = fail;
}

next();

});

</script>
</pre>
</body>
</html>
